fix: change component model to a new file and rename some file
